+++ test/controllers/concerns/require_known_shop_test.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "test_helper"
+
 class RequireKnownShopTest < ActionController::TestCase
   class UnauthenticatedTestController < ActionController::Base
     include ShopifyApp::RequireKnownShop