+++ test/dummy/config/initializers/shopify_app.rb
@@ -12,6 +12,7 @@ class ShopifyAppConfigurer
       config.embedded_app = true
       config.myshopify_domain = "myshopify.com"
       config.api_version = :unstable
+      config.billing = nil
 
       config.shop_session_repository = ShopifyApp::InMemorySessionStore
       config.after_authenticate_job = false